Testing whether we overwrite the vexfile
git-svn-id: svn+ssh://jop91.astron.nl/auto/home/jive_cvs/SOFTC/svn/trunk/sfxc@770 5736e3ea-d426-0410-8aaf-93e336ac3372

+++ utils/produce_html_plotpage.cc
@@ -23,6 +23,18 @@
 #include "output_header.h"
 #include "fringe_info.h"
 
+#include <sys/stat.h>
+
+bool file_exists(char *filename) {
+  struct stat stFileInfo;
+  bool blnReturn;
+  int intStat;
+
+  // Attempt to get the file attributes
+  intStat = stat(filename, &stFileInfo);
+  return intStat;
+}
+
 bool copy_file(char *from, char *to) {
   std::ifstream in(from);
   if (!in.is_open()) return false;
@@ -96,8 +108,11 @@ int main(int argc, char *argv[]) {
     vex_file = new char[strlen(vex_file_temp)+1];
     strcpy(vex_file, vex_file_temp);
     sprintf(to, "%s/%s", output_dir, vex_file);
-    copy_file(from, to);
-
+    if (!file_exists(to)) {
+      copy_file(from, to);
+    } else {
+      std::cout << "Vex-file already exists, not copying" << std::endl;
+    }
   }
 
   if (argc== 4) {
